Prevent int16_t overflow (14857959)
This code is extremely dangerous:

*t += 32 * bonus - *t * abs(bonus) / 256;

The problem is that *t is int16_t, and since signed integer overflow is
undefined in C, the compiler can evaluate this code however it wants, leading
to different outcomes, so the overflow effect (eg. wrap around zero on x86)
will happen at different steps. Regardless, the wrap around zero is not what we
want. We want to do the full calc w/o overflow, apply the cap & floor, and only
then store it into 16-bits for reducing memory usage.

This is actually a functional change, albeit an untestable one, since it only
triggers at very high depths.

+++ src/search.c
@@ -452,7 +452,7 @@ static int search(Worker *worker, const Position *pos, int ply, int depth, int a
         const size_t fuhIdx = stack_move_key(&worker->stack, 1) % NB_FOLLOW_UP;
 
         for (int i = 0; i < quietSearchedCnt; i++) {
-            const int16_t bonus = quietSearched[i] == bestMove ? depth * depth : -1 - depth * depth / 2;
+            const int bonus = quietSearched[i] == bestMove ? depth * depth : -1 - depth * depth / 2;
             const move_t m = quietSearched[i];
             const int from = move_from(m), to = move_to(m);
